requirements.txt
PyStemmer==1.3.0
brewer2mpl==1.4.1
ggplot==0.6.5
ipython==2.3.1
matplotlib==1.4.2
mock==1.0.1
nltk==3.0.0
nose==1.3.4
numpy==1.9.1
pandas==0.15.1
patsy==0.3.0
pyparsing==2.0.3
python-dateutil==2.2
pytz==2014.9
scikit-learn==0.15.2
scipy==0.14.0
six==1.8.0
statsmodels==0.6.0
wsgiref==0.1.2
